cli: move GetDataFromContext and ParseParams to a `helpers` package
