+++ cli/smartcontract/smart_contract.go
@@ -87,11 +87,6 @@ func init() {
 func RuntimeNotify(args []interface{}) {
     runtime.Notify(notificationName, args)
 }`
-	// cosignersSeparator is a special value which is used to distinguish
-	// parameters and cosigners for invoke* commands
-	cosignersSeparator  = "--"
-	arrayStartSeparator = "["
-	arrayEndSeparator   = "]"
 )
 
 // NewCommands returns 'contract' command.
@@ -542,7 +537,7 @@ func invokeInternal(ctx *cli.Context, signAndPush bool) error {
 	paramsStart++
 
 	if len(args) > paramsStart {
-		cosignersOffset, params, err = ParseParams(args[paramsStart:], true)
+		cosignersOffset, params, err = cmdargs.ParseParams(args[paramsStart:], true)
 		if err != nil {
 			return cli.NewExitError(err, 1)
 		}
@@ -620,52 +615,6 @@ func invokeInternal(ctx *cli.Context, signAndPush bool) error {
 	}
 
 	return nil
-}
-
-// ParseParams extracts array of smartcontract.Parameter from the given args and
-// returns the number of handled words, the array itself and an error.
-// `calledFromMain` denotes whether the method was called from the outside or
-// recursively and used to check if cosignersSeparator and closing bracket are
-// allowed to be in `args` sequence.
-func ParseParams(args []string, calledFromMain bool) (int, []smartcontract.Parameter, error) {
-	res := []smartcontract.Parameter{}
-	for k := 0; k < len(args); {
-		s := args[k]
-		switch s {
-		case cosignersSeparator:
-			if calledFromMain {
-				return k + 1, res, nil // `1` to convert index to numWordsRead
-			}
-			return 0, []smartcontract.Parameter{}, errors.New("invalid array syntax: missing closing bracket")
-		case arrayStartSeparator:
-			numWordsRead, array, err := ParseParams(args[k+1:], false)
-			if err != nil {
-				return 0, nil, fmt.Errorf("failed to parse array: %w", err)
-			}
-			res = append(res, smartcontract.Parameter{
-				Type:  smartcontract.ArrayType,
-				Value: array,
-			})
-			k += 1 + numWordsRead // `1` for opening bracket
-		case arrayEndSeparator:
-			if calledFromMain {
-				return 0, nil, errors.New("invalid array syntax: missing opening bracket")
-			}
-			return k + 1, res, nil // `1`to convert index to numWordsRead
-		default:
-			param, err := smartcontract.NewParameterFromString(s)
-			if err != nil {
-				return 0, nil, fmt.Errorf("failed to parse argument #%d: %w", k+1, err)
-			}
-			res = append(res, *param)
-			k++
-		}
-	}
-	if calledFromMain {
-		return len(args), res, nil
-	}
-	return 0, []smartcontract.Parameter{}, errors.New("invalid array syntax: missing closing bracket")
-
 }
 
 func testInvokeScript(ctx *cli.Context) error {
@@ -828,7 +777,7 @@ func contractDeploy(ctx *cli.Context) error {
 		return cli.NewExitError(fmt.Errorf("failed to restore manifest file: %w", err), 1)
 	}
 
-	_, data, extErr := GetDataFromContext(ctx)
+	_, data, extErr := cmdargs.GetDataFromContext(ctx)
 	if extErr != nil {
 		return extErr
 	}
@@ -886,31 +835,6 @@ func contractDeploy(ctx *cli.Context) error {
 	fmt.Fprintf(ctx.App.Writer, "Contract: %s\n", hash.StringLE())
 	fmt.Fprintln(ctx.App.Writer, txHash.StringLE())
 	return nil
-}
-
-// GetDataFromContext returns data parameter from context args.
-func GetDataFromContext(ctx *cli.Context) (int, interface{}, *cli.ExitError) {
-	var (
-		data   interface{}
-		offset int
-		params []smartcontract.Parameter
-		err    error
-	)
-	args := ctx.Args()
-	if args.Present() {
-		offset, params, err = ParseParams(args, true)
-		if err != nil {
-			return offset, nil, cli.NewExitError(fmt.Errorf("unable to parse 'data' parameter: %w", err), 1)
-		}
-		if len(params) != 1 {
-			return offset, nil, cli.NewExitError("'data' should be represented as a single parameter", 1)
-		}
-		data, err = smartcontract.ExpandParameterToEmitable(params[0])
-		if err != nil {
-			return offset, nil, cli.NewExitError(fmt.Sprintf("failed to convert 'data' to emitable type: %s", err.Error()), 1)
-		}
-	}
-	return offset, data, nil
 }
 
 // ParseContractConfig reads contract configuration file (.yaml) and returns unmarshalled ProjectConfig.
